requirements.txt
fastapi
uvicorn
nats-py